Moves shuffle function to helpers
